Separate events from states
