+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "3rd-party/magic_enum"]
+[submodule "external/magic_enum"]
 	path = external/magic_enum
 	url = git@github.com:Neargye/magic_enum.git
