+++ src/game/controller/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory(event)
 add_subdirectory(input)
 add_subdirectory(state)
